+++ CHANGES.md
@@ -7,6 +7,7 @@ v1.1.4 (Month DD, YYYY)
 
 - Fixed detection of encrypted strings that are too short (Issue #52)
 - Fixed a TrueType CMAP decoding bug.
+- Fixed a text rendering issue for Asian text.
 
 
 v1.1.3 (November 15, 2023)